backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
httpx>=0.25
pyyaml>=6.0
sqlglot>=20.0
pandas>=2.0
pydantic>=2.0
python-multipart>=0.0.6

openai>=1.0.0
 
apscheduler>=3.10
croniter>=2.0
tzdata>=2024.1
